Fix BFloat16 numpy conversion error in gpt-interactive
Cast logits to float32 before .numpy() since numpy does not support
BFloat16 scalar type.

+++ gpt-interactive.py
@@ -72,7 +72,7 @@ def _(device, model, mo, np, tokenizer, torch):
     with torch.no_grad():
         _out = model(**_inputs)
     # Logits for the last token position = prediction for the next token
-    raw_logits = _out.logits[0, -1, :].cpu().numpy()
+    raw_logits = _out.logits[0, -1, :].float().cpu().numpy()
 
     # Show top 15 tokens by logit value
     _top_idx = np.argsort(raw_logits)[::-1][:15]
